# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/inspection-tasks": {
      "get": {
        "description": "Inspector sees tasks assigned to them; supervisor sees tasks in their zones + agency. Optional `status` filter.",
        "operationId": "InspectionController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Array of tasks with business/license/report."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List inspection tasks (scoped)",
        "tags": [
          "Inspection"
        ]
      },
      "post": {
        "description": "Assignee must hold the inspector role, share a zone with the business, and match the agency. Rejected (409) if the assignee owns the business (conflict of interest). Generates a sequential `T-YYYY-NNNN` number and notifies the assignee.",
        "operationId": "InspectionController_createTask",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created task."
          },
          "403": {
            "description": "Assignee does not cover the zone."
          },
          "404": {
            "description": "Business or assignee not in scope."
          },
          "409": {
            "description": "Assignee owns the business (conflict of interest)."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create an inspection task (supervisor)",
        "tags": [
          "Inspection"
        ]
      }
    },
    "/api/inspection-tasks/{id}": {
      "get": {
        "description": "Includes business, zone, license, assignee, and the latest report. Returns 404 if outside the caller’s scope.",
        "operationId": "InspectionController_findTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Task uuid",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The task with related data."
          },
          "404": {
            "description": "Not found or out of scope."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get an inspection task (scoped)",
        "tags": [
          "Inspection"
        ]
      }
    },
    "/api/inspection-tasks/{id}/start": {
      "patch": {
        "description": "Transitions ASSIGNED → IN_PROGRESS, sets `startedAt`, and creates a draft report if none exists.",
        "operationId": "InspectionController_startTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Task uuid",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The updated task."
          },
          "422": {
            "description": "Invalid state transition."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Start a task (assignee)",
        "tags": [
          "Inspection"
        ]
      }
    },
    "/api/inspection-tasks/{id}/cancel": {
      "patch": {
        "description": "Only from ASSIGNED or IN_PROGRESS. Requires a reason.",
        "operationId": "InspectionController_cancelTask",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Task uuid",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The cancelled task."
          },
          "422": {
            "description": "Invalid state transition."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Cancel a task (supervisor)",
        "tags": [
          "Inspection"
        ]
      }
    },
    "/api/inspection-reports/{id}": {
      "put": {
        "description": "Editable only while the report is a draft or the task is RETURNED.",
        "operationId": "InspectionController_updateReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Report uuid",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated report."
          },
          "422": {
            "description": "Report is not editable."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a draft report (owner)",
        "tags": [
          "Inspection"
        ]
      }
    },
    "/api/inspection-reports/{id}/evidence": {
      "post": {
        "description": "Multipart upload. Max 10 MB; allowed types: image/jpeg, image/png, application/pdf. Stored privately in MinIO under the report.",
        "operationId": "InspectionController_uploadEvidence",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Report uuid",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": [
                  "file"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The created evidence document."
          },
          "422": {
            "description": "Invalid file or report not editable."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Upload evidence file (owner)",
        "tags": [
          "Inspection"
        ]
      }
    },
    "/api/inspection-reports/{id}/evidence/{docId}": {
      "delete": {
        "description": "Allowed only while the report is a draft or RETURNED.",
        "operationId": "InspectionController_deleteEvidence",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Report uuid",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "docId",
            "required": true,
            "in": "path",
            "description": "Document uuid",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion result."
          },
          "404": {
            "description": "Document not found."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete an evidence file (owner)",
        "tags": [
          "Inspection"
        ]
      }
    },
    "/api/inspection-reports/{id}/submit": {
      "patch": {
        "description": "Requires a non-null result. Sets the report non-draft, transitions the task to PENDING_REVIEW, and notifies supervisors in the zone + agency.",
        "operationId": "InspectionController_submitReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Report uuid",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The submitted report."
          },
          "422": {
            "description": "Result missing or invalid state transition."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Submit a report (owner)",
        "tags": [
          "Inspection"
        ]
      }
    },
    "/api/inspection-reports/{id}/approve": {
      "patch": {
        "description": "Transitions the task to APPROVED. Side effects: FAILED result suspends the license; a PASSED result reactivates a previously suspended one. Notifies the inspector.",
        "operationId": "InspectionController_approveReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Report uuid",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The approved report."
          },
          "404": {
            "description": "Not found or out of scope."
          },
          "422": {
            "description": "Invalid state transition."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Approve a report (supervisor)",
        "tags": [
          "Inspection"
        ]
      }
    },
    "/api/inspection-reports/{id}/return": {
      "patch": {
        "description": "Requires a review comment. Transitions the task to RETURNED and notifies the inspector.",
        "operationId": "InspectionController_returnReport",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Report uuid",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReturnReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The returned report."
          },
          "404": {
            "description": "Not found or out of scope."
          },
          "422": {
            "description": "Invalid state transition."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Return a report for fixes (supervisor)",
        "tags": [
          "Inspection"
        ]
      }
    },
...
  },
  "info": {
    "title": "E-License Verification Platform API",
    "description": "REST API for the Thai e-licensing platform that plugs into the ทางรัฐ\n(Tang Rat) super app. Serves the public app and the admin portal.\n\n**Authentication.** Most endpoints require a Bearer access token\n(RS256 JWT, 15-min TTL). Obtain one from `POST /api/auth/register` or\n`POST /api/auth/login` (PUBLIC), `POST /api/auth/tang-rat`\n(PUBLIC/INSPECTOR/SUPERVISOR), or `POST /api/auth/self` (ADMIN + TOTP).\nClick **Authorize** and paste the `accessToken` to call protected routes.\nSee `docs/AUTHENTICATION.md` for a plain-English walkthrough.\n\n**Scope.** INSPECTOR/SUPERVISOR responses are filtered server-side by the\ncaller's zones and agency; clients cannot widen scope via query params.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "in": "header",
        "description": "Paste the access token returned by an /auth login route."
      }
    },
    "schemas": {
...
      "CreateTaskDto": {
        "type": "object",
        "properties": {}
      },
      "CancelTaskDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateReportDto": {
        "type": "object",
        "properties": {}
      },
      "ReturnReportDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}